+++ components/alert/demo/close.vue
@@ -1,6 +1,9 @@
 <template>
   <Space vertical block>
-    <Alert type="success" showIcon closable>Success Text</Alert>
+    <Alert v-if="visible" type="success" showIcon closable @after-close="visible = false">
+      Success Text
+    </Alert>
+    <Button v-else @click="visible = true">Show Alert</Button>
     <Alert type="info" showIcon closable>Info Text</Alert>
     <Alert type="warning" showIcon closable>Warning Text</Alert>
     <Alert type="error" showIcon closable>Error Text</Alert>
@@ -13,3 +16,7 @@
     />
   </Space>
 </template>
+<script setup lang="ts">
+import { ref } from "vue";
+const visible = ref(true);
+</script>