+++ components/tag/demo/dynamic.vue
@@ -1,6 +1,6 @@
 <template>
   <Space wrap>
-    <Tag color="blue" closeable v-for="t in tags" :key="t">{{ t }}</Tag>
+    <Tag color="blue" closeable v-for="t in tags" :key="t" @after-close="remove(t)">{{ t }}</Tag>
     <Input
       v-if="showInput"
       @blur="add"
@@ -17,7 +17,10 @@ import { Bookmark } from "kui-icons";
 import { nextTick, ref } from "vue";
 const showInput = ref(false);
 const tag = ref("");
-const tags = ["Apple", "Banana", "Cat", "Dog"];
+const tags = ref(["Apple", "Banana", "Cat", "Dog"]);
+const remove = (tag: string) => {
+  tags.value = tags.value.filter((item) => item !== tag);
+};
 const inputRef = ref();
 const show = () => {
   showInput.value = true;
@@ -27,8 +30,8 @@ const show = () => {
 };
 const add = (e: FocusEvent) => {
   let value = (e.target as HTMLInputElement).value.trim();
-  if (value && tags.indexOf(value) === -1) {
-    tags.push(value);
+  if (value && !tags.value.includes(value)) {
+    tags.value.push(value);
   }
   tag.value = "";
   showInput.value = false;